Split discussion picker UI helpers
