+++ service-worker.js
@@ -86,6 +86,7 @@ const APP_SHELL = [
   '/js/chat-discussion-round-state.js',
   '/js/chat-discussion-round-view.js',
   '/js/chat-discussion-state.js',
+  '/js/chat-discussion-picker.js',
   '/js/chat-discussion-ui.js',
   '/js/chat-onboarding.js',
   '/js/chat-empty-state.js',
